add solution JavaSyntax task0709

+++ 1.JavaSyntax/src/com/javarush/task/pro/task07/task0709/Solution.java
@@ -5,8 +5,8 @@ package com.javarush.task.pro.task07.task0709;
 */
 
 public class Solution {
-    long a = (long) (109 + 15);
-    int b = (int) (a * 2);
+    long a = 109 + 15;
+    int b = (int) a * 2;
     short c = (short) (a / b);
     byte d = (byte) (a + b - c);
 }
